% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observables.R
\name{CardiacCycle}
\alias{CardiacCycle}
\alias{cardiac_cycle,}
\alias{cardiac_cycle_observable,}
\alias{cardiac_cycle_observables}
\alias{FirstBeat}
\alias{NormalCycle}
\title{Cardiac Cycle Observables}
\usage{
CardiacCycle()

FirstBeat()

NormalCycle(CardiacCycle)
}
\description{
Observables representing complete cardiac cycles.
}
\section{Functions}{
\itemize{
\item \code{CardiacCycle}: This is the base class to represent
cardiac cycles.

\item \code{FirstBeat}: Class to represent the first heartbeat
in an interpretation, used to break the recursion in the search for the
previous beat.

\item \code{NormalCycle}: This class represents a normal cardiac
cycle, with all its components.
}}

\seealso{
\link{observables}, \link{CardiacRhythm},
\link{QRS}, \link{Deflection}.
}
