% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{construe-package}
\alias{construe}
\alias{construe-package}
\title{R interface to Construe}
\description{
Construe is a knowledge-based abductive framework for time series interpretation.
It provides a knowledge representation model and a set of algorithms for the
interpretation of temporal information, implementing a hypothesize-and-test
cycle guided by an attentional mechanism.
}
\details{
This package provides an R interface to the Construe framework for the interpretation of
electrocardiogram (ECG) signals in terms of the basic waveforms (P, QRS, T) to complex rhythm
patterns (Atrial fibrillation, Bigeminy, Trigeminy, Ventricular flutter/fibrillation, etc.).

See the Construe website at \url{https://github.com/citiususc/construe/} for complete documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/citiususc/construe}
  \item Report bugs at \url{https://github.com/citiususc/construe/issues}
}

}
\author{
\strong{Maintainer}: Constantino A. García \email{constantinoantoniogm@gmail.com}

Authors:
\itemize{
  \item Tomás Teijeiro [copyright holder]
}

}
