% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construe-package.R
\docType{package}
\name{construe}
\alias{construe}
\alias{construe-package}
\title{HTTP Request, Response and URL Parser}
\description{
A dimple and fast HTTP request, response and URL parser based on the C++ 'httpparser' library
by Alex Nekipelov (\url{https://github.com/nekipelov/httpparser})
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/construe}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/construe/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
