% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mit.R, R/utils.R
\name{sp2ms}
\alias{sp2ms}
\alias{ms2sp}
\alias{get_time}
\title{Construe utils}
\usage{
sp2ms(samples)

ms2sp(msec)

get_time()
}
\arguments{
\item{samples}{Number of samples.}

\item{msec}{Time in milliseconds.}
}
\description{
Construe utils
}
\section{Functions}{
\itemize{
\item \code{sp2ms}: Obtains the milliseconds corresponding to a number
of signal samples.

\item \code{ms2sp}: Obtains the number of samples corresponding to a
milliseconds timespan.

\item \code{get_time}: Get time using python module time instead of
Sys.time()
}}

