% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construe_interpretation.R
\name{ConstrueInterpretation}
\alias{ConstrueInterpretation}
\title{Construe interpretation}
\usage{
ConstrueInterpretation()
}
\description{
Creates an object that represents the interpretation entity, which is a consistent
group of abstraction hypotheses combined by the knowledge expressed in abstraction patterns.
It is the basic entity in our search process, and the result of an interpretation process.
}
