% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_construe}
\alias{install_construe}
\alias{install_dependencies}
\title{Install Construe and its dependencies}
\usage{
install_construe(dependencies = TRUE)

install_dependencies()
}
\arguments{
\item{dependencies}{Boolean value indicating if the python dependencies of
Construe should be installed. Default is TRUE.}
}
\description{
Install Construe and its dependencies
}
\section{Functions}{
\itemize{
\item \code{install_construe}: Downloads the python modules required to run
Construe.

\item \code{install_dependencies}: Try to install the python dependencies of
Construe.
}}

