% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observables.R
\name{Observable}
\alias{Observable}
\alias{observable,}
\alias{observables}
\alias{EventObservable}
\title{Observables}
\usage{
Observable()

EventObservable()
}
\description{
Base classes for all the observables.
}
\section{Functions}{
\itemize{
\item \code{Observable}: Base class for all the observables. It has three
attributes: \emph{start}, \emph{end} and \emph{time} (used to represent the
observable as a single event).

\item \code{EventObservable}: This type of observables have only one temporal
variable, to which the two temporal variables of full observables are
referenced.
}}

\seealso{
\link{CardiacCycle}, \link{CardiacRhythm},
\link{QRS}, \link{Deflection}.
}
