% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_request}
\alias{parse_request}
\alias{parse_request_raw}
\title{Parse an HTTP request}
\usage{
parse_request(req, headers_lowercase = TRUE)

parse_request_raw(req, headers_lowercase = TRUE)
}
\arguments{
\item{req}{HTTP request character string}

\item{headers_lowercase}{if \code{TRUE} (the default) names in the \code{headers} data frame
element are converted to lower case}
}
\description{
You can use the non- \verb{_raw} version on input you know for sure is plain text
}
\examples{
paste0(c(
  "GET /uri.cgi HTTP/1.1\r\n",
  "User-Agent: Mozilla/5.0\r\n",
  "Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8\r\n",
  "Host: 127.0.0.1\r\n", "\r\n"
), collapse = "") -> req

res <- parse_request(req)
res <- parse_request_raw(charToRaw(req))
}
