% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_response}
\alias{parse_response}
\alias{parse_response_raw}
\title{Parse an HTTP response}
\usage{
parse_response(resp, headers_lowercase = TRUE)

parse_response_raw(resp, headers_lowercase = TRUE)
}
\arguments{
\item{resp}{HTTP request character string}

\item{headers_lowercase}{if \code{TRUE} (the default) names in the \code{headers} data frame
element are converted to lower case}
}
\description{
You can use the non- \verb{_raw} version on input you know for sure is plain text
}
\examples{
paste0(c(
  "HTTP/1.1 200 OK\r\n",
  "Server: nginx/1.2.1\r\n",
  "Content-Type: text/html\r\n",
  "Content-Length: 8\r\n",
  "Connection: keep-alive\r\n",
  "\r\n",
  "<html />"
), collapse = "") -> resp

res <- parse_response(resp)
res <- parse_response_raw(charToRaw(resp))
}
