% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_url}
\alias{parse_url}
\title{Parse URLs}
\usage{
parse_url(urls)
}
\arguments{
\item{urls}{character vector of URLs}
}
\description{
Parse URLs
}
\examples{
URL <- "http://www.example.com/dir/subdir?param=1&param=2;param\%20=\%20#fragment"
parse_url(URL)
}
