% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_acquisition.R
\name{set_record}
\alias{set_record}
\alias{set_offset}
\alias{set_duration}
\alias{set_tfactor}
\alias{get_tfactor}
\alias{reset_acquisition}
\alias{start_acquisition}
\alias{stop_acquisition}
\alias{get_more_evidence}
\alias{get_acquisition_point}
\alias{get_record_length}
\alias{get_record_name}
\title{Acquisition system}
\usage{
set_record(record, annotator = NULL, physical_units = FALSE)

set_offset(offset)

set_duration(duration)

set_tfactor(tfactor)

get_tfactor()

reset_acquisition()

start_acquisition()

stop_acquisition()

get_more_evidence()

get_acquisition_point()

get_record_length()

get_record_name()
}
\arguments{
\item{record}{Name/Path of the MIT-BIH record containing the input signal.}

\item{annotator}{It can be a string with the name of the annotator, or a list of
annotations with the initial evidence to be interpreted.}

\item{physical_units}{Flag to set whether the signal is in digital or physical units.}

\item{offset}{The offset where the acquisition should start.}

\item{duration}{The duration of the record/fragment. It has to be multiple of
256, and the maximum value currently allowed is 23040.}

\item{tfactor}{Time factor controlling the input speed.}
}
\description{
Interface to control the acquisition system to work with signal records simulating the real-time input.
}
\section{Functions}{
\itemize{
\item \code{set_record}: Sets the record used for input and the initial evidence.

\item \code{set_offset}: Sets an offset in the record from where the
acquisition should start.

\item \code{set_duration}: Sets the duration of the record/fragment to
be interpreted.

\item \code{set_tfactor}: Sets the time factor to control the input speed.

\item \code{get_tfactor}: Obtains the temporal factor controlling the current input speed.

\item \code{reset_acquisition}: Resets the acquisition system.

\item \code{start_acquisition}: Starts the acquisition of signal and evidence.

\item \code{stop_acquisition}: Stops the acquisition of signal and evidence.

\item \code{get_more_evidence}: Obtains the time point, in signal samples, where the
acquisition process is.

\item \code{get_acquisition_point}: Obtains the time point, in signal samples, where the
acquisition process is.

\item \code{get_record_length}: Obtains the length of the record.

\item \code{get_record_name}: Obtains the name of the input reecord.
}}

