% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observables.R
\name{CardiacRhythm}
\alias{CardiacRhythm}
\alias{rhythm,}
\alias{rhythm_observables,}
\alias{rr}
\alias{RR}
\alias{RhythmStart}
\alias{RegularCardiacRhythm}
\alias{SinusRhythm}
\alias{Tachycardia}
\alias{Bradycardia}
\alias{Extrasystole}
\alias{Bigeminy}
\alias{Trigeminy}
\alias{Asystole}
\alias{VentricularFlutter}
\alias{Couplet}
\alias{RhythmBlock}
\alias{AtrialFibrillation}
\title{Rhythm Observables}
\usage{
CardiacRhythm()

RR()

RhythmStart()

RegularCardiacRhythm()

SinusRhythm()

Tachycardia()

Bradycardia()

Extrasystole()

Bigeminy()

Trigeminy()

Asystole()

VentricularFlutter()

Couplet()

RhythmBlock()

AtrialFibrillation()
}
\description{
Observables related with the cardiac rhythm analysis.
}
\section{Functions}{
\itemize{
\item \code{CardiacRhythm}: Class that represents a general and unspecified
cardiac rhythm

\item \code{RR}: Class that represents the RR observable, this
is, the distance between two consecutive ventricular depolarizations.

\item \code{RhythmStart}: Class that represents the start of the first
detected rhythm.

\item \code{RegularCardiacRhythm}: Class that represents a regular rhythm.

\item \code{SinusRhythm}: Class that represents sinus rhythm.

\item \code{Tachycardia}: Class that represents tachycardia rhythm.

\item \code{Bradycardia}: Class that represents bradycardia rhythm.

\item \code{Extrasystole}: This class represents an extrasystole.

\item \code{Bigeminy}: This class represents a bigeminy rhythm.

\item \code{Trigeminy}: This class represents a trigeminy rhythm.

\item \code{Asystole}: Class that represents an asystole (absence of
cardiac activity).

\item \code{VentricularFlutter}: Class that represents a ventricular flutter
rhythm.

\item \code{Couplet}: Class that represents a ventricular couplet
rhythm.

\item \code{RhythmBlock}: Class that represents a rhythm block.

\item \code{AtrialFibrillation}: Class that represents atrial fibrillation.
}}

\seealso{
\link{observables}, \link{CardiacCycle},
\link{QRS}, \link{Deflection}.
}
