% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observables.R
\name{QRS}
\alias{QRS}
\alias{segmentation,}
\alias{segmentation_observables,}
\alias{qrs,}
\alias{t-wave,}
\alias{p-wave}
\alias{PWave}
\alias{TWave}
\alias{QRSShape}
\alias{Noise}
\alias{RPeak}
\alias{Baseline}
\title{Segmentation observables}
\usage{
QRS()

PWave()

TWave()

QRSShape()

Noise()

RPeak()

Baseline()
}
\description{
Observables related with the segmentation of the ECG signal in
components.
}
\section{Functions}{
\itemize{
\item \code{QRS}: Observable that represents a QRS complex
and its shape by lead.

\item \code{PWave}: Observable that represents a P Wave.

\item \code{TWave}: Observable that represents a T Wave.

\item \code{QRSShape}: Class that represents the shape of a QRS
complex in a specific leads. It consists in a sequence of waves, a string tag
abstracting those waves, an amplitude and energy and maximum slope measures,
and a numerical array representing the signal.

\item \code{Noise}: Observable that represents a noisy signal
fragment.

\item \code{RPeak}: Observable that represents a R wave peak.

\item \code{Baseline}: Observable that represents a baseline
observation.
}}

\seealso{
\link{observables}, \link{CardiacCycle},
\link{CardiacRhythm}, \link{Deflection}.
}
