% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observables.R
\name{Deflection}
\alias{Deflection}
\alias{spectrum,}
\alias{spectrum_observables,}
\alias{spectrum_observable,}
\alias{deflection}
\alias{RDeflection}
\title{Spectrum observables}
\usage{
Deflection()

RDeflection()
}
\description{
Observables related with the spectrum analysis of the ECG signal.
}
\section{Functions}{
\itemize{
\item \code{Deflection}: This class represents a signal deviation
consistent with the electrical activity of the cardiac muscle fibers. It is
associated with a certain energy level derived from the wavelet
decomposition/reconstruction  of the signal.

\item \code{RDeflection}: This class represents a signal deviation
consistent with the electrical activity generated in the ventricular
activation. It can be obtained by any external QRS detection algorithm.
}}

\seealso{
\link{observables}, \link{CardiacCycle},
\link{CardiacRhythm}, \link{QRS}.
}
