% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construe_interpretation.R
\name{Construe}
\alias{Construe}
\title{Construe}
\usage{
Construe(root_node, k)
}
\arguments{
\item{root_node}{An object of class \emph{construe_interpretation} (see
\link{ConstrueInterpretation}.)
that determines the root node for the exploration of hypothesis.}

\item{k}{Integer representing the width of the search in the exploration.}
}
\description{
A class that implements the Construe algorithm allowing fine-grained
control of the steps of the algorithm.
}
