% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construe-package.R
\docType{package}
\name{construe}
\alias{construe}
\alias{construe-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/construe}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/construe/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
