% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_file_raw}
\alias{read_file_raw}
\title{Read in a file, fast and raw}
\usage{
read_file_raw(fil)
}
\arguments{
\item{fil}{file to read in (no path expansion is performed)a}
}
\description{
Read in a file, fast and raw
}
